src/main/test/java/com/aoc/y2023/ProblemDay24Test.java
package com.aoc.y2023;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay24Test implements ProblemDayTest {

    private ProblemDay24 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay24();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(13910);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(618534564836937L);
    }

}
